pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bojai"
version = "0.1.0"
description = "BojAI – Modular ML pipelines for everyone"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
authors = [
  { name="Saughmon Boujkian", email="soghmon5@gmail.com" }
]

dependencies = [
  "cryptography==42.0.5",
  "numpy==1.26.4",
  "pandas==2.2.3",
  "Pillow==11.2.1",
  "psutil==6.0.0",
  "PyQt5==5.15.11",
  "PyQt5_sip==12.16.1",
  "pytest==8.3.5",
  "PyYAML==6.0.2",
  "Requests==2.32.3",
  "scikit_learn==1.6.1",
  "torch==2.3.0",
  "transformers==4.41.2"
]

[project.scripts]
bojai = "bojai.cli:main"


[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
